Error is not displayed for functions with no return values. (#2769)
* Error is not displayed for functions with no return values.

Fixes #2749

* fix tests

* update paddings
